updated job sequencing
Hi Shradha DI,
I am Jatin Thakur an ALPHAIT of BELLMAN'S FORD BATCH. While I was doing the class of greedy on apnacollge.in and tried to practice job sequencing problem on GFG
the code didn't work for the test cases then after doing some research I figured out what was the mistake. I tried to contact the hello@apnacollege.in team but didn't get any response. Then I came across this code in your repo and thought it is the best way to contact you 😊 if this account is handled by you. Please see the changes in this code and do the required updates. 
THANK YOU Shradha Di and Aman Bhaiya for being my mentor PROUD TO BE AN ALPHAIT.

At last, Di can I ask you a favor can you please add me to your LinkedIn connection it would be a big thing for me.

LinkedIn- https://www.linkedin.com/in/thakurjatinx/


I sent you the request earlier but about 1 week back I canceled it because there was not any personalized message in it and now I can't send you a connection request for 2 more weeks.
 
THANK AGAIN !!!
Bhaiya and Di